Create Cargo.toml
This Cargo.toml file specifies the package name and version, as well as the authors of the project. It also lists the dependencies that the project needs in order to build, including the serde and serde_json crates for serialization and deserialization, and the sha2 crate for hashing. The dev-dependencies section lists the dependencies that are only needed for development, such as the rand crate for generating random numbers in the unit tests.

You can add additional dependencies as needed by adding them to the dependencies section. For example, if you wanted to use a database to store the ledger data, you could add a database crate such as rusqlite or postgres to the dependencies section.