Create blockchain.rs
In this implementation, the Blockchain struct contains a VecDeque of Blocks, as well as a VecDeque of Transactions that have not yet been included in a block. The new method creates a new instance of the Blockchain struct and adds a "genesis block" to the chain with a proof of 100 and an empty previous hash. The new_block method creates a new block using the transactions in the current transactions queue and the provided proof and previous hash, and adds it to the end of the chain. The new_transaction method adds a new transaction to the current transactions queue, and returns the index of the block that the transaction will be added to. The proof_of_work method implements a simple proof of work algorithm, where it increments the proof until it finds a valid proof (a proof that results in a hash with a leading zero). The valid_proof method verifies that a given proof is valid by hashing it and checking that the resulting hash has a leading zero.